serwer/requirements.txt
Flask
Flask-SQLAlchemy
marshmallow-sqlalchemy
flask-marshmallow
requests
